store plugin settings in separate database table
The configuration of the plugin is now stored in a separate database
table. This table needs to be created first. The SQL script
dbinit/mysql.sql does this for MySQL (schema may change in the future).
To migrate existing settings to the database, the user needs to save the
preferences once. Before doing so, the old address books will not be
used.