caching version for read-only usage
This is a first version that works for a read-only usage of the address
book. Write operations are not yet implemented. Though it seems to work,
it still needs a lot of polish (error checks in particular). Also the
search is not optimal as it does not use the DB to prefilter. Still
haven't quite worked out how all the bits fit together...