detect deleted cards on the server and delete locally
When refreshing contacts from the server, a local cache containing
database id, URI and etag of each locally stored card of the address
book is created.

This cache is used to detect whether a card provided by the server
already exists locally (INSERT vs UPDATE) and also whether the local
copy needs to be updated (compare ETAGs). Previously, a separate DB
query was used to determine this for each individual card.

In addition, whenever a card provided by the server is processed, its
data is removed from the created cache. When the operation is complete,
the cards still in the cache are those that do not exist on the server
anymore, which are then also deleted locally.