don't expose stored carddav passwords to the client in the settings
Prior to this commit, the stored carddav passwords were exposed to the
client in cleartext as presets for the password field in the settings
dialog. Now the password field is empty. A new password will only be set
if the user enters a value in the password field. If the user leaves the
password field empty and changes other settings, the password will not
be changed.