support Apple's X-ABLabel extension for custom labels
Apple's addressbook allows the user to define arbitrary custom labels
for an entry (phone number, mail, address, ...). This commit detects
such labels and stores them in a a DB table. They are then displayed in
the addressbook and become selectable for other entries as well. It is
not possible to define _new_ custom labels from within roundcube though.

The main motivation for this commit is to preserve existing custom
labels when updating a card in roundcube.

X-ABLabel uses group prefixes to link the X-ABLabel attribute to the
corresponding main attribute. For this, I modified vcard.php to store
the group prefix and extended getProperty() by an optional group
parameter that retrieves a property that belongs to the given group.