simple heuristic to infer X-ABShowAs
The showas setting is not implicitly changed on modifying cards
using a simple heuristic: if a organization is given but neither
firstname nor lastname, showas is set to company. If showas was
set to company but no organization is given anymore, showas is
reset to default.

save_data[showas] is now initialized in insert() and update() so
storing a card will not clobber the showas setting.