use If-(None-)Match headers when storing
As recommended in RFC 6352, when creating or updating a card on the
CardDAV server, an appropriate header will be added to avoid lost
updates.

On inserts, the header
 If-None-Match: *
has the server perform the store only if the target URI does not exist.

On updates, the header
 If-Match: $etag
tells the server to perform the store only if the etag of the card on
the server matches the one given in the header, which ensures that the
card was not changed in the meantime.

Both cases are unlikely. If the request fails, the plugin will not retry
or attempt to merge data but just return with an error.