query CardDAV server for addressbook URI
When a new addressbook is added, the CardDAV server is now used to
determine the addressbook URL of the user. For example, to use icloud
it's sufficient to enter https://contacts.icloud.com, and the server
will provide the actual server instance and principal URI used by the
user. Could also find the CardDAV server using a DNS query for the SRV
record, but for that query a separate php package that seems not to be
part of most standard installation would be needed.

The debug functions and cdfopen are now static so they can be used
without having an instance of the carddav_backend at hand.