started adopting the admin presets from trunk
The admin can predefine addressbooks in config.inc.php. Optionally,
the server-related fields can be protected from change. Predefined
addressbooks cannot be deleted (but deactivated) by the user. The
admin may additionally prevent the user to create additional address
books.

I added a new after_login hook which is used to create the preset
addressbooks for the user only once per session. Also moved the
call to the migration function for the old (non-db) settings there.
When the admin deletes presets, they are deleted as well. When the
admin changes settings of a preset, only the admin-only settings
will be updated, since the user may have entered custom values for
the other ones.

Which fields are affected by the 'fixed' attribute (i.e. may not be
changed by the user when this attribute is set) is configurable via
a static array in carddav.php.

The DB schemas were extended with three additional fields:

 - presetname: is set to the name of the preset for preset address
   books. This both marks predefined addressbooks and at the same time
   enables the user to change the name of the addressbook while still
   allowing us to match it to its preset.

 - readonly: attribute to mark address books as readonly. Will probably
   only make sense for presets and not be configurable via the
   interface. This attribute is not yet used by the plugin.

 - fixed: marks preset addressbooks where the user must not change the
   server settings.

Only tried the SQLite schema so far.