bugfix: refresh check did not consider timezone with sqlite backend
The date('now') SQLite function returns the current time in UTC,
whereas MDB's now() returns the current time in the servers timezone.
The former was used as the base for the current time in the check if
a refresh need to occur, but the latter was then used to store the
time of the last refresh in the database. Now the timezone is
considered when checking if a refresh needs to be done.