Disable addressbook search if user provides full URL
When adding a new addressbook, the plugin tries to automatically find
all addressbooks for the user on the given server. This does not work on
all servers (though it should), and it may not be intended that all
addressbooks are added although the user provided the URL to a specific
addressbook.

Now, the addressbook search is disabled if the URL contains a path
specification. It is still enabled if only a pure server URL is
given.

Examples:

http://server.de/ : search enabled
http://server.de/my/address/book: search disabled

This should fix adding address books for snow leopards addressbook
server, albeit only the fully provided URL case works there. This
server returns a HTML page to PROPFIND reqests on the addressbook
home URL (even if the response content type is limited via the
Accept: text/xml header). This was fixed in lion.

For iCloud, davical and lion's addressbook server, I tested that
both of the above cases work as expected.