fall back to addressbook search if no addressbook found under URL
The last commit broke servers that operated on a sub-tree of the
namespace (ownCloud).

The behavior is now to first try if the URL given by the User points
directly to an addressbook collection. If this is the case, this
addressboook is added. If not, the plugin searches for addressbooks,
based on the provided URL.

Tested with iCloud, Snow Leopard (direct addressbook URL only), Lion,
Davical, AND ownCloud ;-)

(also changed the context types from text/xml to application/xml, since
the former is deprecated according to RFC 4918).