list_records always returns firstname and surname
This is a temporary workaround for a bug in roundcube 0.7.2 (reported as
Ticket #1488394) to support roundcube's new displayname setting.
When creating the addressbook listing, roundcube only requests the
displayname from RCMCardDAV, but additionally needs firstname and
lastname to compose the list name.

This bug does not happen with roundcube's builtin addressbook, which
always returns the firstname and surname, even if they were not
requested.