fix preset updates, configurable password scheme
Preset Updates:

The URL can only safely be updated if the preset URL directly refers to
an addressbook. Otherwise, the search result url would be stored with
the addressbook, and overwritten at the next login. Since checking the
URLs upon each login would slow down the login process and cause
annoyance, I currently just added a warning on this issue and removed
'url' from the fixed attributes in the example presets. I have not
generally disabled it because it can be used for direct addressbook
URLs.

Configurable Password Store Scheme:

Upon request of a user, the password encryption is now optional and
the admin can choose how the password is stored in the DB. The options
are:

- plaintext passwords
- base64-coded passwords (not secure, but takes the password out of the
   admin's sight when working with the DB)
- encrypted with the user's IMAP password

The default is base64, because encrypted passwords require the user to
reenter the CardDAV passwords after changing the IMAP password.