fix usage of roundcube password and username in presets.
The $_SESSION was initialized in the login task, the time at which
the admin presets where read. Consequently, presets using the
username or password from the $_SESSION records would not work.

The config.inc.php is now read at a later time, which fixes this.
In addition, %u and %p can be used as placeholders for the roundcube
passwords, which should be more intuitive than accessing the SESSION
record. The placeholders are substituted only right before a connection
to the server is created. Using %p for the password instead of
retrieving the cleartext password from the session has the additional
benefit that only %p will be stored in the database, i.e. the password
of the user will not be stored in the DB at all.

Should fix FS#50 and FS#44.