preprocess_rc_data: check if photo already is base64 encoded
Addresses FS#51

Photos added to cards or updated via the addressbook interface are
provided by roundcube in binary form. If a card is copied from
roundcube's builtin addressbook to the carddav addressbook, however,
the photo is provided in base64 encoded form.

preprocess_rc_savedata now checks if the photo attribute contains
base64 encoded data, and only performs the base64 encoding if this
is not the case