do not push the roundcube adressbook to the end of the list
Addresses FS#52

The delete button is not enabled for the roundcube addressbook
for its non-numeric index in the array. Pushing it to the last
numeric index of the array would word (unless there are further
addressbook plugins that append to the list).

However, I believe that the carddav plugin is not the right
place to affect the global addressbook order. Such functionality
should better be provided by roundcubes addressbook management.