fix: don't delete URL on save for preset addressbooks
For preset addressbooks for that the url attribute was *not* in the list
of fixed attributes, the URL got delete as soon as the user hit save in
the preference pane. The problem was the the URL can never be changed
for preset addressbooks, and hence the URL is never set for preset
addressbooks on save. The URL field is special in this respect, but was
not specially treated.

No the no_override function that respects the URL is used in the save
function to decide whether to store an attribute or not.