avoid an empty displayname for company cards
For company cards with an empty ORG attribute, the set_displayname
function would set the displayname to an empty string. This may cause a
refresh operation to abort, since the resulting records cannot be
inserted into the DB for violating the NOT NULL constraint.

Now, the displayname is left untouched for those cards if the company
name is not set.