make sure that DB URIs contain no %u placeholder
For addressbooks with placeholders in the URL, cards created with the
plugin would have the %u in the URI stored in the DB. This would not
match with the placeholder-free URI provided by the server on a refresh,
and consequently cause these cards to be deleted.

Also fixed the return values of the addvcards() function, which should
return -1, not false, on error. With the correct return value, cards
should not be deleted (as happened above) in the case of a server error.
Also log a warning if errors occurred during the refresh of an
addressbook.