try to create a displayname if not defined
First it is attempted to build the name from first name
and last name, if present (Company cards should already
be correctly identified). If these are not available,
the first email or phone number is used. If none of those
are available, the name is set to 'Unknown Contact', which
may become visible to the user who may change it to something
meaningful.

Should address FS#72:

- don't die on incoming cards that lack FN
- avoid creating such cards ourselves