fix propfind method / Google Contacts support
The PROPFIND method to get the cards in an addressbook collection
(only used as a fallback when sync-collection is not supported by
 the server) was broken. The answer to the PROPFIND includes the
URL of the connection itself. Including the URL of the collection
in the multiget REPORT causes an internal server error on the Google
server.

For google contacts, the use the CardDAV URL
https://google.com/m8/carddav

Support for SRV/TXT records for autodiscoverying this information
may be added in a future release if a suitable DNS class is
available for PHP.