incremental syncs using the sync-collection report
The server-provided sync-token is now remembered and used for subsequent
syncs, so that the server only transmits the cards that changed since
the last synchronization.

The etag is nevertheless checked on incoming cards, which may already
be available locally if they were added by the rcmcarddav plugin (in
that case, the sync-token wont have been updated).