Manually handle permanently moved redirects during bootstrap process
Following redirections is only performed for GET requests by the http
client class (which conforms to the HTTP protocol). During the
autoconfiguration process using the /.well-known/carddav, a redirect
on a PROPFIND request is needed though. cdfopen can now auto-follow
redirections of arbitrary request types, which can be used during the
bootstrap process.