Fix Depth header when looking for an addressbook
Using the Depth:1 header when looking for addressbooks causes
unnecessary traffic and parsing time increases, because the
response will include a list of elements belonging to the
collection.

Asking for D:current-user-principal and C:addressbook-home-set with
Depth:1 also triggers a bug on DAViCal, which recalculates
those values for each item inside the collection. Recalculating
these properties implies some queries to the DB.

This can cause big performance issues if the addressbook contains a
high number of elements.