Add 'RCMC' to all XPath prefixes to avoid conflicts
If any of the the prefixes defined in carddav_common::registerNamespaces() was
already used in the XML document generated by the CardDAV server, SimpleXML
would just ignore it if it was not used to identify the same namespace, so XPath
queries would not use it as expected.

This commit changes XPath prefixes to other ones that aren't likely to be
already used.