Fix group membership removal
Removing a contact from a single group caused it to be removed from every
group the contact was in, at least on the database. vCard were modified as
expected.

This commit adds an $other_conditions argument to the delete_dbrecord() method,
which is used by remove_from_group() to specify which group the contact has to
be removed from.