Change showas handling
When connecting to Google Contacts, dbstore_contact() is called, but insert() and update() are not called.  preprocess_rc_savedata() was therefore not being called, and contacts that have no firstname or surname, but do have an organization were still being shown as individuals.  The call to preprocess_rc_savedata() has therefore been moved from insert() and update() - which both call dbstore_contact() - into dbstore_contact().  Additionally, the defaulting of 'showas' to INDIVIDUAL when save_data is declared was then causing the logic in preprocess_rc_savedata() to never set 'showas' to COMPANY, so this defaulting has been moved into preprocess_rc_savedata() as well.