Rephrase cd_use_categories to be more technical
The previous text uses the term "modern", which is not quite correct and
not really helpful to a user. Actually, the KIND attribute was
standardized as part of vCard4 (we currently use
X-ADDRESSBOOKSERVER-KIND for vCard3), and KIND=group is clearly intented
as the mechanism to represent contact groups. Where as CATEGORIES is a
way to assign Tags to vCards, which not necessarily need to be suited to
represent groups. Some clients use the KIND approach, some use
CATEGORIES, so this is really a matter of interoperability between the
CardDAV clients used by a particular user.