Add a new global "sync_collection_workaround" preference.
Server support for sync-collection is still a bit iffy these days. At
least DAViCal and Radicale have bugs in their support that causes
synchronization to fail. In lieu of a way to detect whether or not
sync-collection works properly, the simplest thing to do is offer
users the ability to skip it.

This commit adds a new global "sync_collection_workaround"
preference. When enabled, the backend will use list_records_propfind()
instead of list_records_sync_collection(). It is off by default.

Github-Issue: 179