Replace substring occurences of username placeholders
Instead of only replacing '%u' and '%l' when they are a literal
match, so a substring replacment of them. This is especially
useful when the Roundcube login doesn't use domain part, but
CardDAV server requires domain part. With this change, it allows
preset configurations to set username to '%u@example.com'.

The documentation offers '%d' as another valid placeholder, but
it wasn't actually being replaced. Fixes this ommission as above.