Fix issue #109
The issue with figuring out the auth method during REPORT calls on Apple CCS
was that the Apple server issues 400 Bad Request when REPORT calls have
no request body *before* it wants to issue a 401 with a WWW-Authenticate
challenge header. To prevent this, simply set the request body before
doing the call, as in this context, it should exist.