Re-add support for CATEGORIES-based groups.
The plugin will show both types of groups (KIND=group vcards and
CATEGORIES property in contact vcards) as groups. Updating a group will
work using the corresponding group type.

The use_categories configuration switch only is meaningful with respect
to the creation of new groups in roundcube, where this switch will
decide whether a CATEGORIES-based group is created or a KIND=group vcard
is created.

A caveat of the CATEGORIES-based group is that empty groups cannot be
represented on the server. As such, a CATEGORIES-type group created in
roundcube will not become visible on the server / to other clients until
the first contact has been added to the group. When a group becomes
empty, it is not automatically cleared from the roundcube database
either. The user will have to manually request the deletion of the
group.

This change requires the adaptation of the database schema to remove
some NOT NULL constraints for the CATEGORIES-type groups. This commit
only contains a tested adaptation for MySQL, and untested one for
PostgreSQL and none at all for SQLite.