Move checkMigrations function to Database class
One step further to isolate rcube_db dependencies to the Database class.
Additionally, removed the reverse dependency of the Database class to
the carddav class by injecting the logger object into the Database
class.

Additionally, tried to simplify the code and improve error handling in
the checkMigrations() function a bit.