Keep values for new addressbook in settings form
This addresses issue #108 - when a user enters settings for a new
addressbook in the form and an error occurs while adding the
addressbook, the values would be gone. This means the user has to start
entering everything from scratch again. Now the values are retained in
the form, allowing the user to correct the error.

This also fixes that the changed use_categories switch was never stored
to the DB, i.e. the setting could not be changed.