Fix: Delete CATEGORY-type groups that become empty during sync
When due to the deletion of contacts during sync a CATEGORY-type becomes
empty, delete it.

Secondly, also delete the contact-to-group assignments as VCard-type
groups and contacts are deleted. This is needed because SQlite3 does per
default not carry out the ON CASCADE constraints, also we attempt to
enable them now.