Treat NULL and empty strings preset keys as no preset
Originally, the presetname used to be NULL for user-created
addressbooks. Within v4, at some point empty strings were inserted
instead of NULL values upon preset creation. This is now reverted to
store NULL values again, but nevertheless an empty string is considered
not a valid preset key.

This is now checked and an error is produced when an admin configures a
preset that has the empty string as preset key.