Use transactions for multi-query operations
Serializable isolation level is used with read-write transactions,
repeatable read isolation level for read-only transactions. Single query
operations are not explicitly wrapped inside a transaction (auto commit
mode).

No operations are currently retried when a transaction failed. For
resync, this is what we want, for other operations, we might have to
reconsider.