Adapt tests framework to allow additional test suites
That is, move stuff for the DB interoperability tests inside
tests/dbinterop, including loading of the roundcube classes. This should
allow to add unit tests that use stubs for the roundcube classes.

Additionally, init the databases with a fresh initial schema before
running the tests.

Includes a first test on whether the transaction isolation works with
the DBs as expected. And found that it does not with MySQL because of a
roundcube bug.