Simple avoidance measure against racing syncs
With the transactions, it is possible - particularly for long running
syncs - that multiple concurrent syncs start because of an impatient
user. In that case, all but one syncs would roll back after having spent
potentially expensive work (communication with CardDAV server). To
reduce the likelihood for this to happen, when a time-triggered sync
starts, we temporarily set the last_updated time such that the next
time-triggered sync would be due 5 minutes from now. If something goes
wrong during the sync, this would "heal" itself in the sense that after
5 minutes, syncs would be attempted again.

This it not done for syncs triggered by change operations (i.e. not
time-based).

Lastly, we now compute times in PHP instead of using database statements
that are not portable across the different DB implementations.