Run DB tests for different DB types in separate phpunit runs
Previously, data providers were used for this, but this is not really
what these are intended for and data providers have some semantics
associated with them that are not favorable and complicate things.

Now, we run a separate phpunit run for each of the database types to
remove this problem from the test code.

This commit also cleans up the Makefile to reduce duplicated statements
by putting the needed DB operations in reusable macros as kind of a DB
abstraction layer, so the targets can be defined once independent of the
DB type.