Intermediate fix for issue 308
Attempts to unify the use of date and time values across DBMS backends.
However, it seems to be more simple to store unix epoch time stamps as
integer in the database, as it would remove the problem of converting
back and forth between unix epoch time and date/time strings also in the
PHP code and simply things overall.

Therefore, this is an intermediate commit to archive work done, but with
intent to replace in the following commits.