Fix: show plugin version on about page
This worked only if a release was installed via composer, not if a
tarball based installation of the plugin was used. Now the plugin
reports the information data itself without relying on roundcube's magic
to determine the info from composer.json and composer.lock files, which
works independent of how the plugin was installed.