Improve testability with PHPUnit
This commit refactors the code such that dependencies are injects,
resulting in fewer static dependencies between our classes. This in turn
enables us to replace the dependencies in tests with stub objects.

To that end, particularly the Database class was converted from a
collection of static methods to a class whose objects are instantiated.
Normally there will only be one instance though. The reason for this is
that phpunit cannot create stubs for static methods, but the Database
class is one that we would need to stub in unit tests that should not
actually access a database.

Generally, the following dependencies are now injected throughout the
code with their instances being created in the plugin main class
(carddav), but may be overridden by test provided objects as optional
argument to the constructor:

- The logger object (previously provided as carddav::$logger)
- The database object (replacing static methods of Database class)

Furthermore all logic dealing with placeholder replacement in
addressbook settings (username, password, url) is now done in carddav,
and also everything showing UI notifications in roundcube.