Fix schematest-mysql for MySQL 8
MySQL 8 seems to dump explicit charset/collation assignments to columns
in the mysqldump even if they are identical to the table's default
charset/collation. Now add them explicitly in the INIT schema so that
the dump after migration is equal again. There seems to be now way to
reset columns to use the default charset of the table.