Cull some dead code
Our approach to multi-value properties when creating a VCard for a
contact updated in roundcube always has been to completely recreate the
value from scratch, disregarding any parameters that could not be
presented to the user in roundcube. A common example is that a property
of a VCard may have multiple labels/subtypes assigned by the TYPE
parameter and/or X-ABLabel. When a contact is stored, all labels but the
one selected by the user in the roundcube contact edit mask are lost.

Some of the code seemed to be not aware of that fact, but the code was
in fact dead. It has now been removed.

This commit also cleans up unused X-ABLabel properties from the VCard.
It should not hurt if these are there, but makes the unittests
difficult.