Fix setting of ORG property in VCards (Fixes #317)
The ORG property of a VCard is structured to reflect an organization and
a hierarchical structure of departments.

Example:
ORG:Company;Division;Department;Group

In roundcube, we only have organization and department as two single
value entries.

Now the above is converted such:
organization: Company
department: Division; Department; Group

The inverse way works as well, where different department levels are
separated by semicolon and optional surrounding whitespace in the
roundcube value.

Furthermore, we remove the handling of department given as an array in
save_data, because roundcube always passes a simple string value only,
and so now das RCMCardDAV.