Unify handling of multi-value attributes
It is now possible to provide a specific handler function to process the
value of a structured attribute (currently ADR), but the rest of the
surrounding code now is shared.

The function will now process all labels found in the roundcube save
data for a multi-value attribute, not only the known ones. It should not
make a practical difference though, as the user is not able to select
labels other than the known ones in the roundcube contact edit mask.