Only set displayname if empty (Fixes #53)
It comes unexpected for users if the formatted name (FN) from a VCard is
overridden by RCMCardDAV because of the X-ABShowAs setting for an
organization. Lets assume that the FN attribute already contains the
intended value. The only situation where we now set a displayname when
parsing a VCard is if it has no FN or an empty FN.

In the other direction, it is also unexpected if a user explicitly sets
a displayname via the corresponding field in roundcube and RCMCardDAV
overrides it. Since we cannot distinguish a displayname composed by
roundcube from one explicitly entered by the user, we leave the
displayname alone except if it is not set.