Simplify DataConversion::toRoundcube
With the delayed download of photos, VCards should be stored in their
original form and not be modified by the data converter anymore
(previously the PHOTO would be inlined if retrieved from URI, or
cropped).

This means the data converter can simply return the saved data and does
not have to indicate anymore whether the vcard was modified.

As a positive side effect, when a photo is not changed but the card is
updated in roundcube, we can send the original PHOTO property to the
server, which in case the PHOTO is referenced by URI will result in a
smaller request body.