Unit tests: new test logic to compare VCards
This comparison is a bit more robust than the simple comparison done by
PHPUnits assertEquals, and allows the tests to execute with
sabre/vobject 3. It still has some assumptions on the order of
properties in the card and the group names of properties, but as long as
this can be worked around in the data set definitions I won't spend the
time to address this in the comparison code.

For now, the older comparision function is still called in addition to
the new one. This is a safety net to detect whether the new code misses
any important property and there to stay until confidence in the new
code is gained.