Fix: spurious error returned when creating VCard on Google
Google assigns a new UID when a VCard is created, causing the search for
the created VCard after resync to fail and displaying a spurious error
in roundcube. It has no other side-effects, i.e. the card is properly
synced and will display, just the feedback on the create operation is an
error.

To resolve this, when we cannot find a card with the corresponding UID
in the DB after resync, we search by URI.