Handle save_data keys without label (Fixes #332)
Roundcube under certain circumstances passed multi-value fields without
a label (e.g. instead of email:home, it uses the key email). One such
situation is when using the "add to addressbook" functionality from the
mail view. Additionally, in this case roundcube does not pass the
addresses as an array, but as a plain string value. This situation was
already handled though.