Fix: Determine full set of contacts when post_search filter is needed
This would only have caused a problem when search is used when the
contacts don't fit on a single page. Roundcube tends to set the page
size to a large number when no paging is desired, so this probably is
not user-visible issue with the way how roundcube uses the interface.

Nevertheless, it's a bug concerning the interface definition.