Use addressbook_export hook to provide exported VCard
Having objects where roundcube does not expect them is error prone. This
shows with the DelayedPhotoLoader in the photo attribute as well, for
which this is actually a workaround. When roundcube prepares a VCard for
export, it casts all objects in save_data values to arrays (expect
DateTime). For DelayedPhotoLoader, this means that not actually the
PHOTO string would be used, but all its members would be stringified
instead. In case an exception is thrown during this process, the export
dies with an uncaught exception. This happened when network connection
to the CardDAV server was not available, since one member of
DelayedPhotoLoader is the AddressbookCollection object, whose toString
method queries the server for the name of the addressbook and throws an
exception if the server cannot be reached.

As a workaround, we place an empty VCard in the vcard attribute first.
Then roundcube will not try to create a VCard from the save_data but
take this vcard string instead. Later on, in the addressbook_export
hook, we overwrite the vcard with the actual data.