Decouple Addressbook options from database row content
This effectively reverts the Addressbook interface to the same as
before, using decoupled psalm data types for the settings.

Before, we simply used the addressbooks DB row attributes. This would
have to be changed after the split. Now we use an array of options for
the addressbook, not linked to the DB columns definition. It is the job
of the caller to provide the options array with the expected content.

The second part of this commit adapts the unit tests for the changed DB
schema, particularly JsonDatabase schema and the test data sets.