Support URL placeholders in Presets for Discovery URL and URL of extra addressbook
The URLs are stored with placeholders replaced in the database (as opposed to username and password, which we store with unsubstituted placeholders).
This means that the username needs to be known by the time we parse config.inc.php, which is specifically not the case when the user logs on. Therefore,
the initialization when the username is not available yet by the time carddav::init() is called is postponed to the afterLogin() hook.