Slight code refactorings to reach full psalm type coverage
Specifically, psalm has issues with callables passed as string to array_map, and isset on multi-dimensional arrays where any level may not be set.
Since it is a rather small amount of code locations that is affected by this, we can change this to use multiple isset statements and anonymous
functions to help psalm here.