Re-discovery for all accounts (Fixes #396)
This is now done in the afterLogin hook after we have initialized the
presets. Upon creation of new accounts from a preset, the discovery is
not directly performed within initPresets anymore, it will directly
occur as the last_discovered time of the new account is at the start of
epoch.