Add --indent in SQLite dump
When adding a column using ADD COLUMN, it will be appended in the same line as the last column of the dumped schema in the normal .schema output. Using --indent., we can achieve a uniform output format that allows us to compare the schema for INIT and the schema after all the migrations without errors caused by missing newline.