Remove PRESET_ATTR_CFGMAP
It's not needed anymore since we have no more collisions between DB columns in accounts and addressbooks. It's not a problem to pass objects to the insert and update functions of AddressbookManager that contain fields of the other object (addressbook vs account), since addressbook manager will check the available columns anyway and also perform checks such as which of those may be updated. Thus we can now simply pass in the preset objects as addressbooks or accounts.