Reduce custom autoloaders where possible
It turns out there is an autoload-dev option in composer that is able to
do most we need. We only need a custom autoloader concerning the loading
of the roundcube classes.

Furthermore, we can also use composer to load the dbmigration PHP
scripts using the classmap feature. I added carddav.php also to the
classmap so psalm is able to find it when being called on a single file
and not the entire project (like PhpStorm does)