Fix: Allow initial creation of template addressbooks for preset account with fixed mandatory fields
The template addressbook for presets is only created when the user saves the preferences of the account in the UI for the first time. When there are fixed fields, these are disabled in the form an no values are retrieved. In such case, we must take the value of the fixed attributes from the preset.