Only consider active, rw addressbooks as candidates for special abooks
The change to the previous behavior is that no error messages will be
logged anymore in case read-only or inactive addressbooks pass the
matchurl/matchname filters and are subsequently sorted out.

This change means the admin will not have no make sure the matchers do
not match read-only (and inactive) addressbooks, as these will not even
be considered in the initial set of candidates.