When updating presets, handle error in querying server-side properties of an addressbook gracefully
When a fixed name setting is used and the template refers to server-side properties of an addressbook, this
querying may fail and would have caused the updating of the preset to be aborted with an error logged. Now
this is handled better in that only the name attribute will not be updated on failure to get the current server
side property values, but the other fixed attributes and the other addressbooks of the server will still be
updated.