Avoid account attributes in a template addressbook
This would happen when the shortcut in AdminSettings::getAddressbookTemplate() was taken to simply return the complete preset as template addressbook. However, this would have caused the addressbook template to override account fields when creating the UI form.
While the latter is fixed in this commit as well, it is not expected to have account attributes inside the template addressbook in the first place, so this is addressed now, too.