Fix unreleased regression: Move/copy cards between carddav addressbooks (#470)
When a card is moved or copied from one CardDAV addressbook to another via the roundcube web interface,
the Addressbook::insert() function will receive a save_data object produced by rcmcarddav's DataConversion
class, which includes a dummy vcard, but also the original vcard object. With the recent change to use a
vcard provided by roundcube upon the import action, this dummy vcard would wrongly be added to the target
addressbook. To avoid this, use the original vcard object that is also included in the save_data when it
has been produced by rcmcarddav.